Observability Dashboards and DBA Resources/Manual Medata Profilers/40-observability/dbsql-query-history-sync/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dbsql-query-history-sync"
dynamic = ["version"]
description = "1> Get dbsql query history. 2> Sync to Delta table."
readme = "README.md"
license = "MIT"
authors = [
    { name = "Nishant Deshpande", email = "nishant.deshpande@databricks.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.7"
dependencies = [
    "requests",
    "dateutils"
]

[project.urls]
Homepage = "https://github.com/databricks/lakehouse-tacklebox/tree/master/40-observability/dbsql-query-history-sync"

[tool.hatch.version]
path = "src/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]
